requirements.txt
streamlit 
openai 
python-dotenv
pandas 
pyperclip
lxml[html_clean]